Refactor integration helpers
